Add Parallel namespace for all parallelised classes
